[Tizen] Change type of the Scene size and remove unused code
* Size type : Vector2->Uint16Pair

+++ dali-extension/devel-api/evas-plugin/evas-plugin.cpp
@@ -27,7 +27,7 @@ namespace Dali
 namespace Extension
 {
 
-EvasPlugin EvasPlugin::New( Evas_Object* parentEvasObject, int width, int height, bool isTranslucent )
+EvasPlugin EvasPlugin::New( Evas_Object* parentEvasObject, uint16_t width, uint16_t height, bool isTranslucent )
 {
   IntrusivePtr< Internal::EvasPlugin > impl = Internal::EvasPlugin::New( parentEvasObject, width, height, isTranslucent );
